+++ packages/blocks/src/CookieBanner/CookieBanner.tsx
@@ -1,11 +1,20 @@
-import React, { FC, useEffect, useState } from 'react';
-import { Alert, Button, Flex, Text } from '@cmpsr/components';
-import { CookieBannerProps } from './types';
+import React, { FC, isValidElement, MouseEvent, useEffect, useState } from 'react';
+import { createContext } from '@chakra-ui/react-utils';
+import { Alert, AlertProps, Button, ButtonProps, Flex, FlexProps, Text } from '@cmpsr/components';
+import { CookieBannerActionsStaticMembers, CookieBannerProps, CookieBannerStaticMembers } from './types';
 
-export const CookieBanner: FC<CookieBannerProps> = ({
+interface CookieBannerContextProps {
+  cookiePolicy: string | undefined;
+  allowAllCookies: () => void;
+  allowRequiredOnly: () => void;
+}
+
+const [CookieBannerProvider, useCookieBannerContext] = createContext<CookieBannerContextProps>({
+  name: 'CookieBannerContext',
+});
+
+export const CookieBanner: FC<CookieBannerProps> & CookieBannerStaticMembers = ({
   children,
-  acceptAllCta = 'Allow all Cookies',
-  acceptRequiredOnlyCta = 'Allow functional only',
   cookieKey = 'cmpsr-cookies-policy',
   cookieVersion = '1.0.0',
   ...rest
@@ -45,37 +54,77 @@ export const CookieBanner: FC<CookieBannerProps> = ({
     setCookiePolicy(newPolicy);
   };
 
-  const handleAllowAllCookies = () => updatePolicy('all');
+  const allowAllCookies = () => updatePolicy('all');
 
-  const handleAllowRequiredOnly = () => updatePolicy('required');
+  const allowRequiredOnly = () => updatePolicy('required');
 
-  const child = typeof children === 'string' ? <Text color="text-secondary">{children}</Text> : children;
+  const areActionsDefined = React.Children.map(
+    children,
+    (child) => isValidElement(child) && child.type === CookieBannerActions
+  ).some(Boolean);
 
-  return shouldShow ? (
-    <Flex
-      direction={{ base: 'column', lg: 'row' }}
-      alignItems="center"
-      bg="background-info"
-      pr="1rem"
-      py="1.5rem"
-      {...rest}
-    >
-      <Alert variant="subtle" pt="0" pr={{ base: '0', lg: '2rem' }} pb={{ base: '2rem', lg: '0' }}>
-        <Alert.Description>{child}</Alert.Description>
-      </Alert>
-       <Flex
-        justifyContent="center"
-        alignItems="center"
-        direction={{base: 'column', lg: 'row'}}
-        gap="1.5rem"
-      >
-        <Button variant="link" onClick={handleAllowRequiredOnly}>
-          {acceptRequiredOnlyCta}
-        </Button>
-        <Button variant="primary" onClick={handleAllowAllCookies}>
-          {acceptAllCta}
-        </Button>
-      </Flex>
-    </Flex>
-  ) : null;
+  return (
+    <CookieBannerProvider value={{ allowAllCookies, allowRequiredOnly, cookiePolicy }}>
+      {shouldShow ? (
+        <Flex
+          direction={{ base: 'column', lg: 'row' }}
+          alignItems="center"
+          bg="background-inform"
+          pr="1rem"
+          py="1.5rem"
+          {...rest}
+        >
+          {children}
+          {!areActionsDefined && (
+            <CookieBanner.Actions>
+              <CookieBanner.Actions.AcceptRequired />
+              <CookieBanner.Actions.AcceptAll />
+            </CookieBanner.Actions>
+          )}
+        </Flex>
+      ) : null}
+    </CookieBannerProvider>
+  );
 };
+
+const CookieBannerContent: FC<AlertProps> = ({ children, ...props }) => (
+  <Alert variant="subtle" pt="0" pr={{ base: '0', lg: '2rem' }} pb={{ base: '2rem', lg: '0' }} {...props}>
+    {typeof children === 'string' ? <Text color="text-secondary">{children}</Text> : children}
+  </Alert>
+);
+CookieBanner.Content = CookieBannerContent;
+
+const CookieBannerActionAcceptAll: FC<ButtonProps> = ({ onClick, children = 'Allow all Cookies', ...props }) => {
+  const { allowAllCookies } = useCookieBannerContext();
+  const handleClick = (event: MouseEvent<HTMLButtonElement>) => {
+    allowAllCookies();
+    onClick?.(event);
+  };
+  return (
+    <Button variant="primary" onClick={handleClick} {...props}>
+      {children}
+    </Button>
+  );
+};
+const CookieBannerActionAcceptRequired: FC<ButtonProps> = ({
+  onClick,
+  children = 'Allow functional only',
+  ...props
+}) => {
+  const { allowRequiredOnly } = useCookieBannerContext();
+  const handleClick = (event: MouseEvent<HTMLButtonElement>) => {
+    allowRequiredOnly();
+    onClick?.(event);
+  };
+  return (
+    <Button variant="link" onClick={handleClick} {...props}>
+      {children}
+    </Button>
+  );
+};
+const CookieBannerActions: FC<FlexProps> & CookieBannerActionsStaticMembers = (props) => (
+  <Flex justifyContent="center" alignItems="center" direction={{ base: 'column', lg: 'row' }} gap="1.5rem" {...props} />
+);
+CookieBannerActions.AcceptAll = CookieBannerActionAcceptAll;
+CookieBannerActions.AcceptRequired = CookieBannerActionAcceptRequired;
+CookieBanner.Actions = CookieBannerActions;