+++ packages/blocks/src/Footer/types.ts
@@ -1,10 +1,20 @@
-import { ReactNode } from 'react';
-import { DividerProps, FlexProps, ImageProps, LinkProps } from '@cmpsr/components';
+import { FC } from 'react';
+import { DividerProps, FlexProps, ImageProps, LinkProps, TextProps } from '@cmpsr/components';
+export { FlexProps as FooterProps } from '@cmpsr/components';
 
-export interface FooterProps extends FlexProps {
-  imageProps: ImageProps;
-  linkGroups: Array<{ title: string; items: Array<LinkProps> }>;
-  copyGroup?: ReactNode;
-  bottomContent?: ReactNode;
-  dividerProps?: DividerProps;
+export interface FooterLinkGroupStaticMember {
+  Title: FC<TextProps>;
+  Link: FC<LinkProps>;
+}
+
+export interface FooterContentStaticMembers {
+  Links: FC<FlexProps>;
+  LinkGroup: FC<FlexProps> & FooterLinkGroupStaticMember;
+  CopyGroup: FC<FlexProps>;
+}
+
+export interface FooterStaticMembers {
+  Logo: FC<ImageProps>;
+  Content: FC<FlexProps> & FooterContentStaticMembers;
+  Bottom: FC<DividerProps>;
 }