+++ packages/components/src/components/primitives/TextPairing/TextPairing.tsx
@@ -1,19 +1,40 @@
-import React, { FC } from 'react';
-import { Stack } from '@chakra-ui/layout';
-import { Text } from '../Text';
-import { TextPairingProps, TextPairingStyles } from '.';
+import React, { createContext, FC, useContext } from 'react';
 import { useStyleConfig } from '@chakra-ui/system';
 import { useResponsiveValue } from '@hooks';
+import { Flex, Text, TextProps, TextVariant } from '@components';
 
-export const TextPairing: FC<TextPairingProps> = ({ variant, labelProps, subLabelProps, textAlign }) => {
+import { TextPairingStyles, TextPairingType } from '.';
+
+export const TextPairingContext = createContext<{
+  labelVariant: TextVariant;
+  subLabelVariant: TextVariant;
+}>({
+  labelVariant: null,
+  subLabelVariant: null,
+});
+
+export const TextPairing: TextPairingType = ({ variant, ...props }) => {
   const responsiveVariant = useResponsiveValue(variant);
-  const styles = useStyleConfig('TextPairing', {
+  const { container, label, subLabel } = useStyleConfig('TextPairing', {
     variant: responsiveVariant,
   }) as TextPairingStyles;
+
   return (
-    <Stack spacing={styles.columnGap} textAlign={textAlign}>
-      <Text {...labelProps} variant={styles.label.variant} />
-      <Text {...subLabelProps} variant={styles.subLabel.variant} />
-    </Stack>
+    <TextPairingContext.Provider
+      value={{
+        labelVariant: label.variant,
+        subLabelVariant: subLabel.variant,
+      }}
+    >
+      <Flex flexDirection="column" {...container} {...props} />
+    </TextPairingContext.Provider>
   );
 };
+
+const TextPairingText: FC<TextProps & { variantProp: string }> = ({ variantProp, ...props }) => {
+  const context = useContext(TextPairingContext);
+  return <Text variant={context[variantProp]} {...props} />;
+};
+
+TextPairing.Label = (props) => <TextPairingText {...props} variantProp="labelVariant" />;
+TextPairing.SubLabel = (props) => <TextPairingText {...props} variantProp="subLabelVariant" />;