+++ packages/components/src/components/primitives/index.ts
@@ -14,6 +14,7 @@ export * from './Icons';
 export * from './Input';
 export * from './Link';
 export * from './Modal';
+export * from './NumberedList';
 export * from './PinInput';
 export * from './ProgressBar';
 export * from './ProgressCircular';
@@ -27,5 +28,5 @@ export * from './Tabs';
 export * from './Tag';
 export * from './Text';
 export * from './TextPairing';
-export * from './Textarea'
+export * from './Textarea';
 export * from './Tooltip';