+++ packages/nextjs-contentful-renderer/src/getPageContent/getPageContent.test.ts
@@ -12,21 +12,14 @@ jest.mock('../utils/getPageId', () => ({
   getPageId: (...params) => mockGetPageId(...params),
 }));
 
-const mockGetPageById = jest.fn();
 const mockGetRouteBySlug = jest.fn();
-const mockConfigNavbar = jest.fn();
-const mockConfigFooter = jest.fn();
 jest.mock('../utils/contentful', () => ({
-  getPageById: (...params) => mockGetPageById(...params),
   getRouteBySlug: (...params) => mockGetRouteBySlug(...params),
 }));
 
-jest.mock('../configNavbar', () => ({
-  configNavbar: (...params) => mockConfigNavbar(...params),
-}));
-
-jest.mock('../configFooter', () => ({
-  configFooter: (...params) => mockConfigFooter(...params),
+const mockGetPageById = jest.fn();
+jest.mock('../utils/getPageById', () => ({
+  getPageById: (...params) => mockGetPageById(...params),
 }));
 
 describe('getPageContent', () => {
@@ -48,6 +41,46 @@ describe('getPageContent', () => {
     expect(mockGetVisitedPageIdFromCookies).toBeCalledTimes(1);
     expect(mockGetVisitedPageIdFromCookies).toBeCalledWith(context, '/');
   });
+  describe('nextjs preview mode', () => {
+    test('should use nextjs preview mode over param', async () => {
+      const preview = true;
+      const context: any = {
+        preview,
+        query: {
+          slug: '/home',
+          preview: !preview,
+        },
+      };
+      await getPageContent(context);
+      expect(mockGetRouteBySlug).toBeCalledTimes(1);
+      expect(mockGetRouteBySlug).toBeCalledWith(expect.anything(), '/home', preview, undefined);
+    });
+    test('should not use nextjs preview mode if false', async () => {
+      const preview = false;
+      const context: any = {
+        preview,
+        query: {
+          slug: '/home',
+          preview: !preview,
+        },
+      };
+      await getPageContent(context);
+      expect(mockGetRouteBySlug).toBeCalledTimes(1);
+      expect(mockGetRouteBySlug).toBeCalledWith(expect.anything(), '/home', !preview, undefined);
+    });
+    test('should not use nextjs preview mode if undefined', async () => {
+      const context: any = {
+        preview: undefined,
+        query: {
+          slug: '/home',
+          preview: true,
+        },
+      };
+      await getPageContent(context);
+      expect(mockGetRouteBySlug).toBeCalledTimes(1);
+      expect(mockGetRouteBySlug).toBeCalledWith(expect.anything(), '/home', true, undefined);
+    });
+  });
   test('should return undefined if no slug match', async () => {
     mockGetRouteBySlug.mockResolvedValueOnce(undefined);
     const pageContent = await getPageContent(fakeContext);
@@ -59,12 +92,12 @@ describe('getPageContent', () => {
     expect(mockGetPageById).toBeCalledTimes(1);
     expect(mockGetPageById).toBeCalledWith(expect.anything(), 'page_id', true);
   });
-  test('should return page content for page id stored in cookies if exists and config navbar', async () => {
+  test('should return page content for page id stored in cookies if exists', async () => {
     mockGetVisitedPageIdFromCookies.mockReturnValueOnce('page_id');
     mockGetPageById.mockResolvedValueOnce(fakePageContent);
     const pageContent = await getPageContent(fakeContext);
     expect(pageContent).toStrictEqual(fakePageContent);
-    expect(mockConfigNavbar).toBeCalledTimes(1);
+    expect(mockGetPageById).toBeCalledTimes(1);
   });
   test('should request route for slug if no content for page in cookies', async () => {
     mockGetVisitedPageIdFromCookies.mockReturnValueOnce('page_id');
@@ -93,20 +126,6 @@ describe('getPageContent', () => {
     await getPageContent(fakeContext);
     expect(mockSetCookie).toBeCalledTimes(1);
     expect(mockSetCookie).toBeCalledWith(fakeContext, '/home', 'page_id');
-  });
-  test('should configure navbar', async () => {
-    mockGetRouteBySlug.mockResolvedValueOnce({});
-    mockGetPageId.mockReturnValueOnce('page_id');
-    mockGetPageById.mockResolvedValueOnce({ ...fakePageContent, navbar: {} });
-    await getPageContent(fakeContext);
-    expect(mockConfigNavbar).toBeCalledTimes(1);
-  });
-  test('should configure footer', async () => {
-    mockGetRouteBySlug.mockResolvedValueOnce({});
-    mockGetPageId.mockReturnValueOnce('page_id');
-    mockGetPageById.mockResolvedValueOnce({ ...fakePageContent, footer: {} });
-    await getPageContent(fakeContext);
-    expect(mockConfigFooter).toBeCalledTimes(1);
   });
   test('should return page content', async () => {
     mockGetRouteBySlug.mockResolvedValueOnce({});