+++ packages/nextjs-contentful-renderer/src/utils/contentful/index.ts
@@ -1,3 +1,4 @@
 export * from './getRouteBySlug';
 export * from './getPageById';
 export * from './getDefaultTheme';
+export * from './getStaticRoutes';